refactor: extract handleInsertKey into insert.ts
